requirements.txt
asgiref==3.11.1
cffi==2.0.0
cryptography==46.0.6
Django==5.2.12
django-cors-headers==4.7.0
djangorestframework==3.17.1
psycopg2-binary==2.9.11
pycparser==3.0
PyJWT==2.12.1
requests==2.32.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
